# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Relay API
  version: 0.1.0
  summary: API-first second-brain backend for long-running AI-assisted work.
servers:
  - url: http://127.0.0.1:8080
    description: Local development
  - url: https://relay.4gly.dev
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque token
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [ok, command, error]
      properties:
        ok:
          type: boolean
          const: false
        command:
          type: string
        error:
          type: object
          required: [code, message, retryable]
          properties:
            code:
              type: string
            message:
              type: string
            retryable:
              type: boolean
            missing_fields:
              type: array
              items:
                type: string
...
    CaptureInput:
      type: object
      properties:
        project:
          type: string
          description: Optional canonical project name. If omitted, the server may infer the project from repo_path for normal flows or use the bound project when it safely matches a project-scoped key.
        repo_path:
          type: string
          description: Optional repo path artifact to attach. When project is omitted, the basename may be used as the inferred project name.
        handoff_path:
          type: string
          description: Optional handoff markdown path to attach.
        design_path:
          type: string
          description: Optional design document path to attach.
        extra_artifacts:
          type: array
          description: Optional additional trusted artifacts such as changed-files manifests, code paths, or PR diffs.
          items:
            $ref: '#/components/schemas/CaptureArtifactInput'
        source:
          type: string
          default: manual
          description: Optional memory source such as chat, markdown, or manual. Defaults to manual when omitted.
        body:
          type: string
          description: Optional raw memory text to store. `note` is accepted as an alias.
        note:
          type: string
          description: Alias for body. Optional raw memory text to store.
        idempotency_key:
          type: string
          description: Optional but recommended write key for retries.
...
    PromoteInput:
      type: object
      properties:
        project:
          type: string
        kind:
          type: string
          enum: [decision, question]
        summary:
          type: string
        reason:
          type: string
        source_note_ids:
          type: array
          items:
            type: string
        source_artifact_ids:
          type: array
          items:
            type: string
        idempotency_key:
          type: string
      required: [project, kind, summary, reason]
...
    IssueAPIKeyInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        scope:
          type: string
          enum: [global, project]
        project:
          type: string
        project_id:
          type: string
      allOf:
        - if:
            properties:
              scope:
                const: project
            required: [scope]
          then:
            anyOf:
              - required: [project]
              - required: [project_id]
...
    RevokeAPIKeyInput:
      type: object
      required: [key_id]
      properties:
        key_id:
          type: string
...
    CaptureResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/CaptureResult'
    PromoteResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/PromoteResult'
    JudgmentTraceWriteResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/JudgmentTraceWriteResult'
    HeuristicProposalCreateResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/HeuristicProposalCreateResult'
...
    IssueAPIKeyResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/IssueAPIKeyResult'
    ListAPIKeysResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/ListAPIKeysResult'
    RevokeAPIKeyResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/RevokeAPIKeyResult'
...
paths:
...
  /v1/api-keys:
    get:
      tags: [memory]
      summary: List issued API keys
      operationId: listApiKeys
      security:
        - BearerAuth: []
      responses:
        '200':
          description: API key list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAPIKeysResponse'
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v1/api-keys/issue:
    post:
      tags: [memory]
      summary: Issue a new reusable API key
      operationId: issueApiKey
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueAPIKeyInput'
      responses:
        '200':
          description: API key issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueAPIKeyResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v1/api-keys/revoke:
    post:
      tags: [memory]
      summary: Revoke a previously issued API key
      operationId: revokeApiKey
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevokeAPIKeyInput'
      responses:
        '200':
          description: API key revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevokeAPIKeyResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: API key not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v1/capture:
    post:
      tags: [memory]
      summary: Capture raw memory and optional artifacts
      operationId: captureMemory
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaptureInput'
      responses:
        '200':
          description: Capture accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaptureResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v1/promote:
    post:
      tags: [memory]
      summary: Promote raw memory into a decision or open question
      operationId: promoteMemory
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromoteInput'
      responses:
        '200':
          description: Promotion accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromoteResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v1/judgment-traces:
    post:
      tags: [style-memory]
      summary: Write a judgment trace for later style-memory evolution
      operationId: writeJudgmentTrace
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JudgmentTraceWriteInput'
      responses:
        '200':
          description: Judgment trace written
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JudgmentTraceWriteResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Project-scoped key does not match requested project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v1/heuristic-proposals:
    post:
      tags: [style-memory]
      summary: Create a pending heuristic proposal
      operationId: createHeuristicProposal
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HeuristicProposalCreateInput'
      responses:
        '200':
          description: Heuristic proposal created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HeuristicProposalCreateResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Project-scoped key does not match requested project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    get:
      tags: [style-memory]
      summary: List heuristic proposals for a project
      description: |
        Returns a paginated list of heuristic proposals for the given project.
        Accepts either an admin bearer token or a relay_session cookie that
        owns the requested project. Cursor is an ISO-8601 created_at keyset.
      operationId: listHeuristicProposals
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: project_id
          required: true
          schema:
            type: string
        - in: query
          name: project
          schema:
            type: string
          description: Optional project name; ignored when project_id is provided.
        - in: query
          name: state
          schema:
            type: string
            enum: [pending, approved, rejected, archived]
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 0
            maximum: 100
            default: 50
        - in: query
          name: cursor
          schema:
            type: string
          description: ISO-8601 timestamp returned in next_cursor of the previous page.
      responses:
        '200':
          description: Heuristic proposals listed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListHeuristicProposalsResponse'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Missing or invalid session cookie / admin bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Caller is not an admin or owner of the requested project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v1/approved-heuristics:
    get:
      tags: [style-memory]
      summary: List approved heuristics for a project
      description: |
        Returns a paginated list of approved heuristics for the given project.
        Accepts either an admin bearer token or a relay_session cookie that
        owns the requested project. Cursor is an ISO-8601 created_at keyset.
      operationId: listApprovedHeuristics
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: project_id
          required: true
          schema:
            type: string
        - in: query
          name: project
          schema:
            type: string
          description: Optional project name; ignored when project_id is provided.
        - in: query
          name: workflow
          schema:
            type: string
        - in: query
          name: artifact_type
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 0
            maximum: 100
            default: 50
        - in: query
          name: cursor
          schema:
            type: string
          description: ISO-8601 timestamp returned in next_cursor of the previous page.
      responses:
        '200':
          description: Approved heuristics listed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListApprovedHeuristicsResponse'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Missing or invalid session cookie / admin bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Caller is not an admin or owner of the requested project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v1/heuristic-proposals/review:
    post:
      tags: [style-memory]
      summary: Approve, reject, or archive a heuristic proposal
      description: |
        Accepts either an admin bearer token or a relay_session cookie owned by
        the project owner. The state transition is conditional on the proposal
        still being in `pending`; concurrent reviewers receive HTTP 409
        `PROPOSAL_ALREADY_RESOLVED`.
      operationId: reviewHeuristicProposal
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HeuristicProposalReviewInput'
      responses:
        '200':
          description: Heuristic proposal reviewed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HeuristicProposalReviewResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Missing or invalid session cookie / admin bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Caller is not an admin or owner of the proposal's project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Heuristic proposal not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Proposal already resolved by another reviewer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v1/approved-heuristics/update:
    post:
      tags: [style-memory]
      summary: Update approved heuristic state
      description: |
        Accepts either an admin bearer token or a relay_session cookie owned by
        the project owner.
      operationId: updateApprovedHeuristic
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovedHeuristicUpdateInput'
      responses:
        '200':
          description: Approved heuristic updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovedHeuristicUpdateResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Missing or invalid session cookie / admin bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Caller is not an admin or owner of the heuristic's project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Approved heuristic not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v1/packets/build:
    post:
      tags: [memory]
      summary: Build an agent-ready packet from stored memory
      operationId: buildPacket
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PacketBuildInput'
      responses:
        '200':
          description: Packet built
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PacketBuildResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /v1/onboarding:
    post:
      tags: [onboarding]
      summary: Validate and store the user's Anthropic API key
      description: |
        Validates the supplied Anthropic API key against api.anthropic.com /v1/models
        and, on success, envelope-encrypts it under the active KEK and persists it for
        the cookie-session user. Idempotent: re-onboarding rotates the ciphertext and
        the row's aad_salt while preserving default_project_id.

        Trust promise: encrypted at rest. Transmitted only to Anthropic for validation,
        never stored or logged in plaintext.

        Auth: cookie session. Admin bearer is rejected (no user identity to bind the
        row to). Strict decoder: any unknown field (including the legacy relay_url
        field removed in D3) returns 400 UNKNOWN_JSON_FIELD.
      operationId: onboardingComplete
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OnboardingCompleteInput'
      responses:
        '200':
          description: Onboarding completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnboardingCompleteResponse'
        '400':
          description: |
            Validation failed. error.code is one of:
            MISSING_REQUIRED_FIELDS, UNKNOWN_JSON_FIELD,
            INVALID_ANTHROPIC_KEY, ANTHROPIC_QUOTA.
            steps[] is always populated so Frame 3 can render chips.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Missing session cookie or admin-only bearer (admin has no user identity to bind to)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Server misconfiguration (e.g. RELAY_DATA_ENCRYPTION_KEY missing)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '502':
          description: Could not reach api.anthropic.com (ANTHROPIC_UNREACHABLE)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags: [onboarding]
      summary: Forget the user's Anthropic API key (D5 trust escape hatch)
      description: |
        NULLs the encrypted key + nonce + aad_salt + onboarding_completed_at.
        The row remains so that prior packets bound to default_project_id still
        resolve, and so re-onboarding can rotate to a fresh ciphertext +
        aad_salt without losing project history.
      operationId: onboardingDeleteKey
      responses:
        '200':
          description: Key forgotten
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnboardingDeleteResponse'
        '401':
          description: Missing session cookie or admin-only bearer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: User has not completed onboarding
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'